토큰 재발급 및 검증 수정

+++ src/main/java/com/limited/sales/token/TokenServiceImpl.java
@@ -1,6 +1,6 @@
 package com.limited.sales.token;
 
-import com.limited.sales.config.GsonSingleton;
+import com.limited.sales.exception.sub.LoginException;
 import com.limited.sales.exception.sub.TokenException;
 import com.limited.sales.redis.RedisService;
 import com.limited.sales.user.vo.User;
@@ -9,56 +9,45 @@ import com.limited.sales.utils.JwtUtils;
 import com.limited.sales.utils.JwtValidationUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
-import javax.validation.constraints.NotNull;
-import java.util.Optional;
+import javax.validation.constraints.NotBlank;
 
 @Service
+@Validated
 @RequiredArgsConstructor
-public final class TokenServiceImpl implements TokenService {
-  private final RedisService redisService;
+public class TokenServiceImpl implements TokenService {
+  private final RedisService<String> redisService;
 
   @Override
-  public String reissue(@NotNull final User user) {
-    final String refreshToken = redisService.getValue(user.getEmail());
+  public TokenVo reissue(final User user) {
+    if (!JwtValidationUtils.isRefreshTokenValid(redisService.getValue(user.getEmail()))) {
+      throw new TokenException("리프레쉬 토큰이 정상적이지 않습니다.");
+    }
 
-    Optional.ofNullable(refreshToken)
-        .filter(JwtValidationUtils::isRefreshTokenValid)
-        .orElseThrow(
-            () -> {
-              throw new TokenException("리프레쉬 토큰이 정상적이지 않습니다.");
-            });
-
-    final String token = JwtProperties.TOKEN_PREFIX + JwtUtils.createAccessToken(user);
-
-    final TokenVo tokenVo = TokenVo.builder().accessToken(token).build();
-    return GsonSingleton.getGson().toJson(tokenVo);
+    return TokenVo.builder()
+        .accessToken(JwtProperties.TOKEN_PREFIX + JwtUtils.createAccessToken(user))
+        .build();
   }
 
   @Override
-  public void deleteRefreshToken(@NotNull final User user) {
-    final String refreshToken = redisService.getValue(user.getEmail());
-
-    Optional.ofNullable(refreshToken)
-        .filter(JwtValidationUtils::isRefreshTokenValid)
-        .orElseThrow(
-            () -> {
-              throw new TokenException("재로그인이 필요합니다.");
-            });
+  public void deleteRefreshToken(final User user) {
+    if (!JwtValidationUtils.isRefreshTokenValid(redisService.getValue(user.getEmail()))) {
+      throw new LoginException("재 로그인이 필요합니다.");
+    }
 
     redisService.deleteValue(user.getEmail());
   }
 
   @Override
-  public void blacklistAccessToken(@NotNull final User user, @NotNull final String authorization) {
-    Optional.ofNullable(authorization)
-        .filter(JwtValidationUtils::hasValidJwtToken)
-        .orElseThrow(
-            () -> {
-              throw new TokenException("엑세스 토큰이 존재하지 않거나 올바르지 않습니다.");
-            });
+  public void blacklistAccessToken(
+      final User user, @NotBlank(message = "토큰이 존재하지 않습니다.") final String authorization) {
+    if (JwtValidationUtils.hasValidJwtTokenNull(authorization)) {
+      throw new TokenException("엑세스 토큰이 올바르지 않습니다.");
+    }
 
-    final String replacePrefixToken = JwtUtils.replaceTokenPrefix(authorization);
-    redisService.setValue(user.getEmail() + JwtProperties.BLACKLIST_POSTFIX, replacePrefixToken);
+    redisService.setValue(
+        user.getEmail() + JwtProperties.BLACKLIST_POSTFIX,
+        JwtUtils.replaceTokenPrefix(authorization));
   }
 }
